sql/task_2_solution/requirements.txt
decorator==5.1.1
filelock==3.8.0
idna==3.4
mysql-connector==2.2.9
platformdirs==2.5.2
uritools==4.0.0
urlextract==1.6.0